% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildSigMatrix}
\alias{buildSigMatrix}
\title{buildSigMatrix}
\usage{
buildSigMatrix(markers.lst, avg.expr, min.size = 10, max.size = 200)
}
\arguments{
\item{markers.lst}{A list of markers for cell types.}

\item{avg.expr}{Average gene expression of reference.}

\item{min.size}{Minimum size of each subset. Default: 10.}

\item{max.size}{Maximum size of each subset. Default: 200.}
}
\value{
Signature matrix.
}
\description{
buildSigMatrix
}
