% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAnnot.R
\name{cellAnnot}
\alias{cellAnnot}
\title{cellAnnot}
\usage{
cellAnnot(
  seurat_obj,
  db.type = "CellMarker",
  Tissue,
  Species = "Human",
  Cancer = TRUE
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{db.type}{scType, CellMarker, CellTaxonomy, scMayoMap or VITA CytBase.}

\item{Tissue}{}

\item{Species}{Human or Mouse.}

\item{Cancer}{TRUE or FALSE}
}
\description{
cellAnnot
}
