% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createSeuratObjAndQC}
\alias{createSeuratObjAndQC}
\title{createSeuratObjAndQC}
\usage{
createSeuratObjAndQC(ref.expr, ref.anno)
}
\arguments{
\item{ref.expr}{Reference gene expression profile.}

\item{ref.anno}{Cell types corrsponding to the single cells in the reference.}
}
\value{
Seurat object.
}
\description{
createSeuratObjAndQC
}
