% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomScore.R
\name{getPvalues}
\alias{getPvalues}
\title{getPvalues}
\usage{
getPvalues(obs.score, rand.score)
}
\arguments{
\item{obj.seu}{Seurat object, which need to be annotated.}

\item{ref.obj}{Reference Seurat object.}
}
\value{
Cell Type Score from Permutation Test.
}
\description{
getPvalues
}
