% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomScore.R
\name{getRandomScore}
\alias{getRandomScore}
\title{getRandomScore}
\usage{
getRandomScore(
  obj.seu,
  lr.model,
  sig.mat,
  sig.features,
  cluster.col,
  sample.times = 100
)
}
\arguments{
\item{obj.seu}{Seurat object, which need to be annotated.}

\item{ref.obj}{Reference Seurat object.}
}
\value{
Cell Type Score from Permutation Test.
}
\description{
getRandomScore
}
