% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identSeedGenes}
\alias{identSeedGenes}
\title{identSeedGenes}
\usage{
identSeedGenes(avg.expr, factor.size = 0.1, count = 10)
}
\arguments{
\item{avg.expr}{Reference gene expression profile.}

\item{factor.size}{Factor size to scale the weight. Default: 0.1.}

\item{count}{Number of candidate query genes. Default: 10.}
}
\value{
A list of seed genes for cell types.
}
\description{
identSeedGenes
}
