% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoMethods.R
\name{lrPredModel}
\alias{lrPredModel}
\title{lrPredModel}
\usage{
lrPredModel(obj.ref, obj.seu, down.sample = 100)
}
\arguments{
\item{obj.ref}{Seurat object of reference gene expression profile.}

\item{obj.seu}{Seurat object of target gene expression profile.}
}
\value{
A list of predicted results.
}
\description{
lrPredModel
}
