% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoMethods.R
\name{mergeModelScore}
\alias{mergeModelScore}
\title{mergeModelScore}
\usage{
mergeModelScore(coef.vals, lr.pred.score, ref.obj, obj.seu, cluster.col)
}
\value{
A list of predicted results.
}
\description{
mergeModelScore
}
