% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoMethods.R
\name{rlmDewcon}
\alias{rlmDewcon}
\title{rlmDewcon}
\usage{
rlmDewcon(ref, bulk, weight = NULL)
}
\arguments{
\item{ref}{Cell type-specific gene reference expression profiles,the reference profile is the mean value of gene expression in the cell type}

\item{bulk}{Gene expression profiling,bulk is the gene expression profile after clustering.}

\item{weight}{Assign the weight of each genes, default: NULL.}
}
\value{
A data.frame of coefficients for cell types.
}
\description{
rlmDewcon
}
