% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{scAnnotation}
\alias{scAnnotation}
\title{scAnnotation}
\usage{
scAnnotation(
  seurat_obj,
  annot.method = "marker",
  datasetREF = "BlueprintEncode",
  db.type = "CellMarker",
  Tissue,
  Species = "Human",
  Cancer = TRUE
)
}
\arguments{
\item{annot.method}{dataset or marker}

\item{db.type}{If annot.method = 'marker',Options are:
`scType`, `CellMarker`, `CellTaxonomy`,`scMayoMap`, `VITA CytBase`.
Only `scType` not support Mouse.}

\item{Cancer}{}

\item{dataset_ref}{If annot.method = 'dataset', built-in
datasets will be used. Options are: `BlueprintEncode`,
`DICE`, `HPCA`, `ImmGen`, `MonacoImmune`, `MouseRNAseq`,
`Hematopoiesis`.}
}
\description{
scAnnotation
}
\examples{
\dontrun{
library(scAnnotation)
result <- scAnnotation(example_data,annot.method='dataset')
result_2 <- scAnnotation(example_data,annot.method='marker')
}

}
