% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scoring.R
\name{Scoring}
\alias{Scoring}
\title{Scoring}
\usage{
Scoring(expr, gs)
}
\arguments{
\item{expr}{input scRNA-seq matrix (rounames - genes, column names - cells)}

\item{gs}{list of gene sets positively expressed in the cell type}
}
\description{
Scoring
}
