% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers.R
\name{searchMarkersByASeurat}
\alias{searchMarkersByASeurat}
\title{searchMarkersBySeurat}
\usage{
searchMarkersByASeurat(obj, ...)
}
\arguments{
\item{obj}{Seurat object.}
}
\value{
A list of markers for cell types.
}
\description{
Search marker genes of each cell type using Seurat method.
}
