% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{searchMarkersByROC}
\alias{searchMarkersByROC}
\title{searchMarkersByROC}
\usage{
searchMarkersByROC(obj, ...)
}
\arguments{
\item{obj}{Seurat object.}
}
\value{
A list of markers for cell types.
}
\description{
Search marker genes of each cell type using ROC method.
}
