% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectFeatures}
\alias{selectFeatures}
\title{selectFeatures}
\usage{
selectFeatures(ref.obj, markers, sig.mat, subset.size = NULL, rand.seed = 1234)
}
\arguments{
\item{ref.obj}{Seurat object.}

\item{markers}{Cell type-specific genes.}

\item{sig.mat}{Signature matrix.}

\item{subset.size}{Marker size of each cell type. Default: NULL.}

\item{rand.seed}{Set random seed. Default: 1234.}
}
\value{
signature gene list
}
\description{
Apply Lasso method to filter the features of logistic regression.
}
