% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{signature_gene}
\alias{signature_gene}
\title{signature_gene}
\usage{
signature_gene(ref.obj, markers)
}
\arguments{
\item{ref.obj}{Seurat object}

\item{markers}{Cell type-specific genes}
}
\value{
signature gene list
}
\description{
Apply Lasso method to filter the features of logistic regression
}
