% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subanno.R
\name{subanno}
\alias{subanno}
\title{Extract cell type information from annotation results}
\usage{
subanno(annotations)
}
\arguments{
\item{annotations}{The annotation result data box processed by the cell_annotation function must contain the 'cluster' and 'content' columns}
}
\value{
Returns a data box with two columns, cluster and celltype:
\itemize{
  \item cluster
  \item celltype - Cell type after analysis (Format: Type [subtype])
}
}
\description{
Extract cell type information from annotation results
}
