% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downSamplSeurat}
\alias{downSamplSeurat}
\title{downSamplSeurat}
\usage{
downSamplSeurat(obj, seed = 123, percent = 0.5)
}
\arguments{
\item{obj}{Seurat object}

\item{seed}{set random seed, default: 123}

\item{percent}{Sampling ratio}
}
\value{
Seurat sub object
}
\description{
downSamplSeurat
}
