% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnno.R
\name{scAnno}
\alias{scAnno}
\title{scAnno}
\usage{
scAnno(
  query = obj.seu,
  ref.expr = ref.expr,
  ref.anno = ref.anno,
  save.markers = NULL,
  cluster.col = "seurat_clusters",
  factor.size = 0.1,
  pvalue.cut = 0.01,
  method = c("co.exp", "diff.exp"),
  seed.num = 10,
  subset.size = NULL,
  redo.markers = FALSE,
  gene.anno = data(gene.anno),
  permut.num = 100,
  show.plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{ref.expr}{Reference gene expression profile.}

\item{ref.anno}{Cell type information of reference profile, corresponding to the above \code{ref.expr}.}

\item{save.markers}{Specified the filename of makers need to be saved.}

\item{cluster.col}{Column name of clusters to be annotated in meta.data slot of query Seurat object. Default: seurat_clusters.}

\item{factor.size}{Factor size for scaling the weight of gene expression. Default: 0.1.}

\item{method}{Specified method for selecting candidate marker genes, co.exp \link{co-expression with seed genes} or diff.exp \link{differential expressed genes by Seurat method}. Default: co.exp.}

\item{seed.num}{Number of seed genes of each cell type for recognizing candidate markers, only used when method = 'co.exp'. Default: 10.}

\item{subset.size}{Marker size of each cell type from reference profile. Default: NULL.}

\item{redo.markers}{Re-search candidate markers or not. Default: FALSE.}

\item{gene.anno}{Gene annotation data.frame. Default: data(gene.anno).}

\item{permut.num}{Number of permutations for estimating p-values of annotations. Default: 100.}

\item{show.plot}{Show annotated results or not. Default: TRUE.}

\item{verbose}{Show running messages or not. Default: TRUE.}

\item{obj.seu}{Seurat object, which need to be annotated.}
}
\value{
Annotated Seurat object.
}
\description{
scAnno
}
