% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnno.R
\name{scAnno}
\alias{scAnno}
\title{scAnno}
\usage{
scAnno(
  query = obj.seu,
  ref.expr = ref.expr,
  ref.anno = ref.anno,
  show.plot = TRUE,
  factor.size = 0.1
)
}
\arguments{
\item{ref.expr}{Reference gene expression profile.}

\item{ref.anno}{Cell type information of reference profile, corresponding to the above \code{ref.expr}.}

\item{show.plot}{Show annotated results or not. Default: TRUE.}

\item{obj.seu}{Seurat object, which need to be annotated.}
}
\value{
Annotated Seurat object.
}
\description{
scAnno
}
