% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanno.R
\name{scanno}
\alias{scanno}
\title{AI single-cell annotation}
\usage{
scanno(
  markers,
  selected_clusters = NULL,
  api_key = Sys.getenv("API_KEY"),
  api_url = Sys.getenv("API_URL"),
  model = "deepseek-r1-250120",
  gene_number = 100,
  workers = 6,
  background = NULL,
  max_retries = 3,
  timeout = 200,
  retry_delay = 1,
  temperature = 0.3,
  max_tokens = 8190
)
}
\arguments{
\item{markers}{The results of FindAllMarkers must include the cluster, p-val, avg.log2FC, and gene columns.}

\item{selected_clusters}{Select the cluster (character vector) for analysis}

\item{api_key}{API key, read by default from the environment variable API_KEY}

\item{api_url}{API request address, read by default from the environment variable API-URL}

\item{model}{Model name used}

\item{gene_number}{The first n genes were selected for annotation}

\item{workers}{Number of parallel working processes}

\item{background}{Sample background description information}

\item{max_retries}{Maximum retry times, default 3 times}

\item{timeout}{Request timeout (seconds)}

\item{retry_delay}{Wait time for retry (seconds)}

\item{temperature}{API temperature parameter, default 0.3}

\item{max_tokens}{The maximum number of tokens for the API is 8190 by default}
}
\description{
This function utilizes an AI big language model to automatically annotate cell clusters in single-cell sequencing data.
}
\examples{
\dontrun{
markers=read.csv("data/all_DEG.csv",row.names = 1)
Sys.setenv(API_URL = "https://api.nbchat.site/v1/chat/completions")
Sys.setenv(API_KEY = "sk-xxxxxxxxx")
anno <- scanno(
  markers,
  selected_clusters = c("0","1"),
  background = "Human peripheral blood single-cell data",
  workers = 6
)
subanno <- subanno(anno)
}
}
